API/requirements.txt
python-dotenv===1.0.0

# mongodb
pymongo==4.6.1

# API-END point
fastapi==0.109.2
fastapi-cors
uvicorn[standard]==0.17.*
python-multipart==0.0.9
pydantic
requests-toolbelt